pyproject.toml
[tool.poetry]
name = "pet_sitter"
version = "0.1.0"
description = ""
authors = ["laurencecnerual <47619680+laurencecnerual@users.noreply.github.com>"]
readme = "README.md"
packages = [
  { include = "pet_sitter" }
]

[tool.poetry.scripts]
start = "pet_sitter.main:start"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.7"}
python-dotenv = "^1.0.1"
pydantic = "^2.10.0"



[tool.aerich]
tortoise_orm = "pet_sitter.main.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
